sprint2/step2 - reducer moved to redux
