+++ src/components/burger-constructor/components/burger-constructor-order/burger-constructor-order.tsx
@@ -1,13 +1,16 @@
-import React, { useCallback, useContext } from 'react';
+import React, { useCallback } from 'react';
 import styles from './burger-constructor-order.module.css';
 import { Button } from '@ya.praktikum/react-developer-burger-ui-components';
 import { MoneyCounter } from '../../../common/money-counter/money-counter';
-import { AppContext } from '../../../../services/AppContext';
-import { IBurgerActionType } from '../../../../model/IBurgerAction';
+import { IBurgerActionType } from '../../../../services/actions';
+import { useDispatch, useSelector } from 'react-redux';
+import { IAppState } from '../../../../model/IAppState';
 
 
 export function BurgerConstructorOrder() {
-	const { state, dispatch } = useContext(AppContext);
+	const dispatch = useDispatch();
+	const state = useSelector(state => ({ ...state })) as IAppState;
+
 	const { sum } = state;
 
 	const onOrderButtonClick = useCallback(() => {
