--- src/services/reducers/index.ts
@@ -1,7 +1,0 @@
-import { combineReducers } from 'redux';
-import { apiReducer } from './api';
-
-
-export const rootReducer = combineReducers({
-	api: apiReducer
-});
